#285 - mu: more image support
