+++ src/runtime/mu-ld/image.rs
@@ -2,11 +2,114 @@
 //  SPDX-License-Identifier: MIT
 
 //! runtime loader
-use crate::{bump_allocator::BumpAllocatorImageBuilder, reader::Reader, writer::Writer};
+#[allow(unused_imports)]
+use crate::{
+    mu_core::{Condition, Env, Result, Tag},
+    reader::Reader,
+    writer::Writer,
+};
 
-pub struct Image {}
+use json::{self, object};
+
+pub struct Image {
+    pub image: Vec<u8>,
+    pub alloc_map: Vec<u8>,
+    pub free_map: Vec<u8>,
+    pub write_barrier: Vec<u8>,
+}
+
+pub struct ImageBuilder {
+    image: Option<Vec<u8>>,
+    alloc_map: Option<Vec<u8>>,
+    free_map: Option<Vec<u8>>,
+    write_barrier: Option<Vec<u8>>,
+}
+
+impl ImageBuilder {
+    pub fn new() -> Self {
+        Self {
+            image: None,
+            alloc_map: None,
+            free_map: None,
+            write_barrier: None,
+        }
+    }
+
+    pub fn image(&mut self, image: Vec<u8>) -> &mut Self {
+        self.image = Some(image);
+        self
+    }
+
+    pub fn alloc_map(&mut self, map: Vec<Vec<u8>>) -> &mut Self {
+        let mut vec: Vec<u8> = vec![];
+
+        for context in map {
+            vec.extend(context)
+        }
+
+        self.alloc_map = Some(vec);
+        self
+    }
+
+    pub fn free_map(&mut self, map: Vec<Vec<u8>>) -> &mut Self {
+        let mut vec: Vec<u8> = vec![];
+
+        for context in map {
+            vec.extend(context)
+        }
+
+        self.free_map = Some(vec);
+        self
+    }
+
+    pub fn write_barrier(&mut self, write_barrier: usize) -> &mut Self {
+        self.write_barrier = Some(write_barrier.to_le_bytes().to_vec());
+        self
+    }
+
+    pub fn build(&self) -> Image {
+        Image {
+            image: match &self.image {
+                Some(vec) => vec.to_vec(),
+                None => vec![],
+            },
+            alloc_map: match &self.alloc_map {
+                Some(vec) => vec.to_vec(),
+                None => vec![],
+            },
+            free_map: match &self.free_map {
+                Some(vec) => vec.to_vec(),
+                None => vec![],
+            },
+            write_barrier: match &self.write_barrier {
+                Some(vec) => vec.to_vec(),
+                None => vec![0, 0, 0, 0, 0, 0, 0, 0, 0],
+            },
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {}
 
 impl Image {
+    pub fn to_json(&self) -> Option<String> {
+        let image = object! {
+            image: self.image.clone(),
+            alloc_map: self.alloc_map.clone(),
+            free_map: self.free_map.clone(),
+            write_barrier: self.write_barrier.clone(),
+        };
+
+        Some(image.dump())
+    }
+
+    /*
+    pub fn from_json(_json: String) -> Option<Self> {
+    None
+    }
+     */
+
     pub fn dump(path: &str) {
         let reader = Reader::with_reader(path).unwrap();
 
@@ -49,14 +152,14 @@ impl Image {
     }
 
     pub fn output(path: &str) {
-        let bump_allocator = BumpAllocatorImageBuilder::new()
+        let image = ImageBuilder::new()
             .image(vec![1, 2, 3])
             .alloc_map(vec![vec![4], vec![5], vec![6]])
             .free_map(vec![vec![7], vec![8], vec![9]])
             .write_barrier(0)
             .build();
 
-        let writer = Writer::with_writer(path, bump_allocator).unwrap();
+        let writer = Writer::with_writer(path, image).unwrap();
 
         writer.write().unwrap()
     }
